requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
alembic==1.13.1
python-multipart==0.0.6
httpx==0.25.2
schedule==1.2.1
loguru==0.7.2
redis==5.0.1
celery==5.3.4